requirements.txt
streamlit>=1.37
anthropic>=0.40
python-dotenv>=1.0
